+++ XOSLLOAD/IO/Fs.cpp
@@ -22,7 +22,7 @@ CFileSystem::~CFileSystem()
 	delete Disk;
 }
 
-int CFileSystem::Mount(int Drive, unsigned long StartSector)
+int CFileSystem::Mount(int Drive, unsigned long long StartSector)
 {
 	return Disk->Map(Drive,StartSector);
 }
